Added reference_python; renamed reference.py to reference_jython.
Fixed up ib/aux/__init__.py to implement all of the expected behavior
of translated code.

Corrected imports and __init__ method in the configuration module for
EReader.py.  Updated ib/ext/EReader.py also included in this checkin.

Changed the method sorter to just an assignment in the configuration
module for EClientSocket.py.



git-svn-id: http://ibpy.googlecode.com/svn/branches/ast@171 3e6b72d3-b629-0410-ad1b-3b4ab856adc4