Fix for unregisterAll() Silent Failure
As per https://github.com/blampe/IbPy/issues/14, unregisterAll() silently fails. The problem is that messageType strings must be passed to self.unregister() as the 2nd+ argument in order for the unregistering process to work properly, but what's being passed in as the second+ argument is *self.messageTypes.values(), which evaluates to 1-tuples and 2-tuples of the Message objects themselves. This causes unregisterAll() to fail silently. The fix passes in messageType strings, which makes unregisterAll() work properly.